@@ -26,7 +26,7 @@
         }
     },
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.0",
         "twig/twig": "^2.0|^3.0",
         "symfony/property-access": "^4.4|^5.0",
         "symfony/dependency-injection": "^4.4|^5.0"
