@@ -11,14 +11,16 @@
 
 namespace Symfony\UX\TwigComponent\DependencyInjection;
 
+use Symfony\Component\DependencyInjection\Argument\AbstractArgument;
 use Symfony\Component\DependencyInjection\Argument\ServiceLocatorArgument;
-use Symfony\Component\DependencyInjection\Argument\TaggedIteratorArgument;
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Reference;
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 use Symfony\UX\TwigComponent\ComponentFactory;
-use Symfony\UX\TwigComponent\ComponentInterface;
 use Symfony\UX\TwigComponent\ComponentRenderer;
+use Symfony\UX\TwigComponent\DependencyInjection\Compiler\TwigComponentPass;
 use Symfony\UX\TwigComponent\Twig\ComponentExtension;
 use Symfony\UX\TwigComponent\Twig\ComponentRuntime;
 
@@ -31,31 +33,36 @@ final class TwigComponentExtension extends Extension
 {
     public function load(array $configs, ContainerBuilder $container): void
     {
-        $container->registerForAutoconfiguration(ComponentInterface::class)
-            ->addTag('twig.component')
-        ;
+        if (method_exists($container, 'registerAttributeForAutoconfiguration')) {
+            $container->registerAttributeForAutoconfiguration(
+                AsTwigComponent::class,
+                static function (ChildDefinition $definition) {
+                    $definition->addTag('twig.component');
+                }
+            );
+        }
 
-        $container->register(ComponentFactory::class)
+        $container->register('ux.twig_component.component_factory', ComponentFactory::class)
             ->setArguments([
-                new ServiceLocatorArgument(new TaggedIteratorArgument('twig.component', null, 'getComponentName')),
+                class_exists(AbstractArgument::class) ? new AbstractArgument(sprintf('Added in %s.', TwigComponentPass::class)) : new ServiceLocatorArgument(),
                 new Reference('property_accessor'),
             ])
         ;
 
-        $container->register(ComponentRenderer::class)
+        $container->register('ux.twig_component.component_renderer', ComponentRenderer::class)
             ->setArguments([
                 new Reference('twig'),
             ])
         ;
 
-        $container->register(ComponentExtension::class)
+        $container->register('ux.twig_component.twig.component_extension', ComponentExtension::class)
             ->addTag('twig.extension')
         ;
 
-        $container->register(ComponentRuntime::class)
+        $container->register('ux.twig_component.twig.component_runtime', ComponentRuntime::class)
             ->setArguments([
-                new Reference(ComponentFactory::class),
-                new Reference(ComponentRenderer::class),
+                new Reference('ux.twig_component.component_factory'),
+                new Reference('ux.twig_component.component_renderer'),
             ])
             ->addTag('twig.runtime')
         ;
